search-website-functions-v4/api-v4-with-directories/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-common==1.1.28
azure-core==1.26.0
azure-functions==1.17.0
azure-search-documents==11.3.0
certifi==2022.12.7
chardet==5.0.0
charset-normalizer==2.1.1
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.12